Five

+++ 02Vydry/Pages/Delete.cshtml.cs
@@ -27,7 +27,6 @@ namespace _02Vydry.Pages
 
         [BindProperty]
         public Vydra Vydra { get; set; }
-        public IList<Vydra> Mothers { get; set; }
 
         public async Task<IActionResult> OnGetAsync(int? id)
         {
@@ -38,7 +37,7 @@ namespace _02Vydry.Pages
 
             Vydra = await _context.Vydras
                 .Include(v => v.Mother)
-                .Include(v => v.Place).ThenInclude(l => l.Location)
+                .Include(v => v.Place).ThenInclude(p => p.Location)
                 .Include(v => v.founder).AsNoTracking().FirstOrDefaultAsync(m => m.TattooID == id);
 
             if (Vydra == null)
@@ -60,21 +59,19 @@ namespace _02Vydry.Pages
                 return NotFound();
             }
 
-            Vydra = await _context.Vydras.FindAsync(id);
-
-            Mothers = _context.Vydras.Include(v => v.Mother).AsNoTracking().ToList<Vydra>();
+            Vydra = await _context.Vydras.Include(v => v.Children).AsNoTracking().FirstOrDefaultAsync(m => m.TattooID == id);
 
             if (Vydra != null)
             {
-                foreach (var item in Mothers)
+                if (Vydra.Children.Count == 0)
                 {
-                    if (item.Mother.TattooID == id)
-                    {
-                        return RedirectToPage("./DeleteError");
-                    }
+                    _context.Vydras.Remove(Vydra);
+                    await _context.SaveChangesAsync();
                 }
-                _context.Vydras.Remove(Vydra);
-                await _context.SaveChangesAsync();
+                else
+                {
+                    return RedirectToPage("./DeleteError");
+                }
             }
 
             return RedirectToPage("./Index");
